--- src/nest/ai/common/services/i-ai-translator-service.ts
@@ -1,13 +1,0 @@
-import { SourceLanguage } from '../../../../utils/language';
-import { FilePathInfo } from '@/types/cache';
-
-export interface AiTranslateParam {
-  sourceTexts: string[];
-  sourceLanguage: SourceLanguage;
-  fileInfo?: FilePathInfo;
-  maxOutputTokenCount: number;
-  requestsPerMinute: number;
-  apiKey: string;
-  promptPresetContent: string;
-  useThinking: boolean;
-}
