+++ src/nest/ai/services/ai-prompt-converter.service.ts
